+++ driver-core/CHANGELOG.rst
@@ -19,9 +19,9 @@ CHANGELOG
 Merged from 2.0 branch:
 
 - [bug] Handle null pool in PooledConnection.release (JAVA-449)
-- [bug] Prevent concurrent retries of the same query (JAVA-449)
 - [improvement] Defunct connection on request timeout (JAVA-425)
 - [improvement] Try next host when we get a SERVER_ERROR (JAVA-426)
+- [bug] Handle race between query timeout and completion (JAVA-449, JAVA-460, JAVA-471)
 
 
 2.1.1:
@@ -89,9 +89,9 @@ Merged from 2.0 branch: everything up to 2.0.3 (included), and the following.
 ------
 
 - [bug] Handle null pool in PooledConnection.release (JAVA-449)
-- [bug] Prevent concurrent retries of the same query (JAVA-449)
 - [improvement] Defunct connection on request timeout (JAVA-425)
 - [improvement] Try next host when we get a SERVER_ERROR (JAVA-426)
+- [bug] Handle race between query timeout and completion (JAVA-449, JAVA-460, JAVA-471)
 
 
 2.0.6: